Moved classes related to transfering to transfer_service.h and updated style.
