--- include/TransferOptions.hpp
@@ -1,19 +1,0 @@
-/*
- * TransferOptions.hpp
- * Andrew Mikalsen
- * 7/7/20
- */
-
-#ifndef TRANSFER_OPTIONS_HPP_INCLUDED
-#define TRANSFER_OPTIONS_HPP_INCLUDED
-
-namespace ods {
-
-    // TODO: define
-    class TransferOptions {
-        public:
-            virtual ~TransferOptions() = 0;
-    };
-}
-
-#endif // TRANSFER_OPTIONS_HPP_INCLUDED
