--- include/TransferStatus.hpp
@@ -1,19 +1,0 @@
-/*
- * TransferStatus.hpp
- * Andrew Mikalsen
- * 7/7/20
- */
-
-#ifndef TRANSFER_STATUS_HPP_INCLUDED
-#define TRANSFER_STATUS_HPP_INCLUDED
-
-namespace ods {
-
-    // TODO: define
-    class TransferStatus {
-        public:
-            virtual ~TransferStatus() = 0;
-    };
-}
-
-#endif // TRANSFER_STATUS_HPP_INCLUDED
